requirements.txt
icon.ico
selenium
customtkinter
Pillow
pypdf4
threading
re
os
io
requests
